src/python/stockminer/requirements.txt
datedelta==1.4
holidays==0.13
numpy==1.21.5
pandas==1.4.2
pandas_datareader==0.10.0
python_dateutil==2.8.2
requests==2.25.1
requests_html==0.10.0
scipy==1.8.0
setuptools==59.6.0
ta==0.10.1
XlsxWriter==3.0.3
xlwt==1.3.0
yahoo_earnings_calendar==0.6.0
yfinance==0.1.70
